pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "marketing-mcp"
version = "0.1.0"
description = "AI-powered marketing data MCP server"
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
    "google-ads>=25.0",
    "google-api-python-client>=2.0",
    "google-auth>=2.0",
    "google-analytics-data>=0.18",
    "pytrends>=4.9",
    "praw>=7.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-cov>=5.0",
    "ruff>=0.5",
]

[project.scripts]
marketing-mcp = "marketing_mcp.server:main"

[tool.hatch.build.targets.wheel]
packages = ["src/marketing_mcp"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
src = ["src"]
